Cleanup, app context, restore functionality
