+++ src/graph/graph_editor_egui/viewport_split.rs
@@ -1,6 +1,6 @@
 use egui::*;
 
-use super::editor_state::EditorState;
+use super::editor_state::GraphEditorState;
 
 #[derive(Clone)]
 pub enum ViewportSplitKind {
