+++ src/mesh/debug_viz.rs
@@ -3,7 +3,7 @@ use crate::prelude::*;
 
 /// Loads a wavefront obj file from a given path and returns a Rend3 mesh
 /// @CopyPaste from wavefront_obj.rs
-pub fn load_obj_mesh(path: &str) -> r3::Mesh {
+fn load_obj_mesh(path: &str) -> r3::Mesh {
     use std::fs::File;
     use std::io::BufReader;
     use wavefront_rs::obj;
@@ -46,6 +46,81 @@ pub struct DebugMeshes {
 }
 
 impl DebugMeshes {
+    pub fn new(renderer: &r3::Renderer) -> DebugMeshes {
+        let cylinder = renderer.add_mesh(load_obj_mesh("./assets/debug/arrow.obj"));
+        let sphere = renderer.add_mesh(load_obj_mesh("./assets/debug/icosphere.obj"));
+        let base_material = renderer.add_material(r3::PbrMaterial {
+            albedo: r3::AlbedoComponent::Value(Vec4::ONE),
+            ..Default::default()
+        });
+
+        let dbg_meshes = DebugMeshes {
+            cylinder,
+            sphere,
+            base_material,
+            color_material_cache: HashMap::new(),
+        };
+
+        // Forget about 'em! We don't want to store these and they must be alive for
+        // the whole application's lifetime
+        Box::leak(Box::new(dbg_meshes.cylinder.clone()));
+        Box::leak(Box::new(dbg_meshes.sphere.clone()));
+        Box::leak(Box::new(dbg_meshes.base_material.clone()));
+
+        dbg_meshes
+    }
+
+    pub fn add_halfedge_debug(&mut self, render_ctx: &mut RenderContext, mesh: &HalfEdgeMesh) {
+        const VERTEX_THICKNESS: f32 = 0.07;
+        const EDGE_THICKNESS: f32 = 0.05;
+        const HALFEDGE_SEPARATION: f32 = 0.03;
+
+        for (v_id, vertex) in mesh.iter_vertices() {
+            let material = self.get_material(&render_ctx.renderer, mesh.vertex_debug_mark(v_id));
+            render_ctx.add_object(r3::Object {
+                mesh_kind: r3::ObjectMeshKind::Static(self.sphere.clone()),
+                material,
+                transform: glam::Mat4::from_translation(vertex.position)
+                    * glam::Mat4::from_scale(Vec3::ONE * VERTEX_THICKNESS),
+            });
+        }
+
+        for (h, _) in mesh.iter_halfedges() {
+            let face_centroid = mesh
+                .at_halfedge(h)
+                .face()
+                .try_end()
+                .map(|face| mesh.face_vertex_average(face));
+
+            let (src, dst) = mesh.at_halfedge(h).src_dst_pair().unwrap();
+            let src_pos = mesh.vertex_position(src);
+            let dst_pos = mesh.vertex_position(dst);
+
+            let midpoint = (src_pos + dst_pos) / 2.0;
+            let delta = dst_pos - src_pos;
+            let delta_dir = delta.normalize();
+
+            let orientation = Quat::from_rotation_arc(Vec3::Y, delta_dir);
+            let orientation = Mat4::from_rotation_translation(orientation, Vec3::ZERO);
+
+            let material = self.get_material(&render_ctx.renderer, mesh.halfedge_debug_mark(h));
+
+            let towards_face = if let Ok(centroid) = face_centroid {
+                (centroid - midpoint).normalize() * HALFEDGE_SEPARATION
+            } else {
+                Vec3::ZERO
+            };
+
+            render_ctx.add_object(r3::Object {
+                mesh_kind: r3::ObjectMeshKind::Static(self.cylinder.clone()),
+                material,
+                transform: Mat4::from_translation(midpoint + towards_face)
+                    * orientation
+                    * Mat4::from_scale(Vec3::new(EDGE_THICKNESS, delta.length(), EDGE_THICKNESS)),
+            });
+        }
+    }
+
     pub fn get_color_material(
         &mut self,
         renderer: &r3::Renderer,
@@ -78,85 +153,5 @@ impl DebugMeshes {
         } else {
             self.base_material.clone()
         }
-    }
-}
-
-pub fn add_debug_meshes(renderer: &r3::Renderer) -> DebugMeshes {
-    let cylinder = renderer.add_mesh(load_obj_mesh("./assets/debug/arrow.obj"));
-    let sphere = renderer.add_mesh(load_obj_mesh("./assets/debug/icosphere.obj"));
-    let base_material = renderer.add_material(r3::PbrMaterial {
-        albedo: r3::AlbedoComponent::Value(Vec4::ONE),
-        ..Default::default()
-    });
-
-    let dbg_meshes = DebugMeshes {
-        cylinder,
-        sphere,
-        base_material,
-        color_material_cache: HashMap::new(),
-    };
-
-    // Forget about 'em! We don't want to store these and they must be alive for
-    // the whole application's lifetime
-    Box::leak(Box::new(dbg_meshes.cylinder.clone()));
-    Box::leak(Box::new(dbg_meshes.sphere.clone()));
-    Box::leak(Box::new(dbg_meshes.base_material.clone()));
-
-    dbg_meshes
-}
-
-pub fn add_halfedge_debug(
-    render_ctx: &mut RenderContext,
-    debug_meshes: &mut DebugMeshes,
-    mesh: &HalfEdgeMesh,
-) {
-    const VERTEX_THICKNESS: f32 = 0.07;
-    const EDGE_THICKNESS: f32 = 0.05;
-    const HALFEDGE_SEPARATION: f32 = 0.03;
-
-    for (v_id, vertex) in mesh.iter_vertices() {
-        let material =
-            debug_meshes.get_material(&render_ctx.renderer, mesh.vertex_debug_mark(v_id));
-        render_ctx.add_object(r3::Object {
-            mesh_kind: r3::ObjectMeshKind::Static(debug_meshes.sphere.clone()),
-            material,
-            transform: glam::Mat4::from_translation(vertex.position)
-                * glam::Mat4::from_scale(Vec3::ONE * VERTEX_THICKNESS),
-        });
-    }
-
-    for (h, _) in mesh.iter_halfedges() {
-        let face_centroid = mesh
-            .at_halfedge(h)
-            .face()
-            .try_end()
-            .map(|face| mesh.face_vertex_average(face));
-
-        let (src, dst) = mesh.at_halfedge(h).src_dst_pair().unwrap();
-        let src_pos = mesh.vertex_position(src);
-        let dst_pos = mesh.vertex_position(dst);
-
-        let midpoint = (src_pos + dst_pos) / 2.0;
-        let delta = dst_pos - src_pos;
-        let delta_dir = delta.normalize();
-
-        let orientation = Quat::from_rotation_arc(Vec3::Y, delta_dir);
-        let orientation = Mat4::from_rotation_translation(orientation, Vec3::ZERO);
-
-        let material = debug_meshes.get_material(&render_ctx.renderer, mesh.halfedge_debug_mark(h));
-
-        let towards_face = if let Ok(centroid) = face_centroid {
-            (centroid - midpoint).normalize() * HALFEDGE_SEPARATION
-        } else {
-            Vec3::ZERO
-        };
-
-        render_ctx.add_object(r3::Object {
-            mesh_kind: r3::ObjectMeshKind::Static(debug_meshes.cylinder.clone()),
-            material,
-            transform: Mat4::from_translation(midpoint + towards_face)
-                * orientation
-                * Mat4::from_scale(Vec3::new(EDGE_THICKNESS, delta.length(), EDGE_THICKNESS)),
-        });
     }
 }
